Correções para convivencia com o novo db

+++ src/main/java/com/labengsoftware/engsoftware/controller/SneakerController.java
@@ -1,6 +1,7 @@
 package com.labengsoftware.engsoftware.controller;
 
 
+import java.math.BigInteger;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -106,7 +107,7 @@ public class SneakerController {
 	}
 	
 	@PatchMapping("/{sneaker_id}")
-	public SneakerResponseDTO updateSneaker(@PathVariable Integer sneaker_id, UpdateSneakerRequestDTO request) {
+	public SneakerResponseDTO updateSneaker(@PathVariable BigInteger sneaker_id, UpdateSneakerRequestDTO request) {
 
 		log.info("Atualizando tenis do usuario");
 
@@ -116,7 +117,7 @@ public class SneakerController {
 	}
 	
 	@DeleteMapping("/{sneaker_id}")
-	public SneakerResponseDTO deleteSneaker(@PathVariable Integer sneaker_id) {
+	public SneakerResponseDTO deleteSneaker(@PathVariable BigInteger sneaker_id) {
 
 		log.info("Deletando tenis");
